Move client-communication structure (API) into production code.
